order-service/requirements.txt
Django==5.0.*
djangorestframework
psycopg[binary]
redis
django-redis
python-dotenv
django-filter
faker
djangorestframework-simplejwt[crypto]
django-environ
coverage
drf-spectacular 
requests
pika

django-prometheus
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-django
opentelemetry-instrumentation-requests
opentelemetry-exporter-otlp